+++ hutool-crypto/src/main/java/cn/hutool/crypto/digest/Digester.java
@@ -5,9 +5,12 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
+import java.security.Provider;
 
 import cn.hutool.core.io.FileUtil;
+import cn.hutool.core.io.IORuntimeException;
 import cn.hutool.core.io.IoUtil;
+import cn.hutool.core.util.ArrayUtil;
 import cn.hutool.core.util.CharsetUtil;
 import cn.hutool.core.util.HexUtil;
 import cn.hutool.core.util.StrUtil;
@@ -26,6 +29,8 @@ public class Digester extends BouncyCastleSupport {
 	private MessageDigest digest;
 	/** 盐值 */
 	protected byte[] salt;
+	/** 加盐位置，既将盐值字符串放置在数据的index数，默认0 */
+	protected int saltPosition;
 	/** 散列次数 */
 	protected int digestCount;
 
@@ -42,11 +47,32 @@ public class Digester extends BouncyCastleSupport {
 	/**
 	 * 构造
 	 * 
-	 * @param algorithm 算法
-	 * @since 3.3.0
+	 * @param algorithm 算法枚举
 	 */
 	public Digester(String algorithm) {
-		init(algorithm);
+		this(algorithm, null);
+	}
+
+	/**
+	 * 构造
+	 * 
+	 * @param algorithm 算法
+	 * @param provider 算法提供者，null表示JDK默认，可以引入Bouncy Castle等来提供更多算法支持
+	 * @since 4.5.1
+	 */
+	public Digester(DigestAlgorithm algorithm, Provider provider) {
+		init(algorithm.getValue(), provider);
+	}
+
+	/**
+	 * 构造
+	 * 
+	 * @param algorithm 算法
+	 * @param provider 算法提供者，null表示JDK默认，可以引入Bouncy Castle等来提供更多算法支持
+	 * @since 4.5.1
+	 */
+	public Digester(String algorithm, Provider provider) {
+		init(algorithm, provider);
 	}
 	// ------------------------------------------------------------------------------------------- Constructor end
 
@@ -54,12 +80,13 @@ public class Digester extends BouncyCastleSupport {
 	 * 初始化
 	 * 
 	 * @param algorithm 算法
+	 * @param provider 算法提供者，null表示JDK默认，可以引入Bouncy Castle等来提供更多算法支持
 	 * @return {@link Digester}
 	 * @throws CryptoException Cause by IOException
 	 */
-	public Digester init(String algorithm) {
+	public Digester init(String algorithm, Provider provider) {
 		try {
-			digest = MessageDigest.getInstance(algorithm);
+			digest = (null == provider) ? MessageDigest.getInstance(algorithm) : MessageDigest.getInstance(algorithm, provider);
 		} catch (NoSuchAlgorithmException e) {
 			throw new CryptoException(e);
 		}
@@ -79,6 +106,30 @@ public class Digester extends BouncyCastleSupport {
 	}
 
 	/**
+	 * 设置加盐的位置，只有盐值存在时有效<br>
+	 * 加盐的位置指盐位于数据byte数组中的位置，例如：
+	 * 
+	 * <pre>
+	 * data: 0123456
+	 * </pre>
+	 * 
+	 * 则当saltPosition = 2时，盐位于data的1和2中间，既第二个空隙，既：
+	 * 
+	 * <pre>
+	 * data: 01[salt]23456
+	 * </pre>
+	 * 
+	 * 
+	 * @param salt 盐值
+	 * @return this
+	 * @since 4.4.3
+	 */
+	public Digester setSaltPosition(int saltPosition) {
+		this.saltPosition = saltPosition;
+		return this;
+	}
+
+	/**
 	 * 设置重复计算摘要值次数
 	 * 
 	 * @param digestCount 摘要值次数
@@ -86,6 +137,17 @@ public class Digester extends BouncyCastleSupport {
 	 */
 	public Digester setDigestCount(int digestCount) {
 		this.digestCount = digestCount;
+		return this;
+	}
+
+	/**
+	 * 重置{@link MessageDigest}
+	 * 
+	 * @return this
+	 * @since 4.5.1
+	 */
+	public Digester reset() {
+		this.digest.reset();
 		return this;
 	}
 
@@ -168,32 +230,25 @@ public class Digester extends BouncyCastleSupport {
 	 * @return 摘要bytes
 	 */
 	public byte[] digest(byte[] data) {
-		if (null != this.salt) {
-			digest.update(this.salt);
+		byte[] result = null;
+		if (this.saltPosition <= 0) {
+			// 加盐在开头，自动忽略空盐值
+			result = doDigest(this.salt, data);
+		} else if (this.saltPosition >= data.length) {
+			// 加盐在末尾，自动忽略空盐值
+			result = doDigest(data, this.salt);
+		} else if (ArrayUtil.isNotEmpty(this.salt)) {
+			// 加盐在中间
+			this.digest.update(data, 0, this.saltPosition);
+			this.digest.update(this.salt);
+			this.digest.update(data, this.saltPosition, data.length - this.saltPosition);
+			result = this.digest.digest();
+		} else {
+			// 无加盐
+			result = doDigest(data);
 		}
-		int digestCount = Math.max(1, this.digestCount);
-		byte[] result = data;
-		for (int i = 0; i < digestCount; i++) {
-			result = doDigest(result);
-		}
-		return result;
-	}
 
-	/**
-	 * 生成摘要
-	 * 
-	 * @param data 数据bytes
-	 * @return 摘要bytes
-	 * @since 4.4.3
-	 */
-	private byte[] doDigest(byte[] data) {
-		byte[] result;
-		try {
-			result = digest.digest(data);
-		} finally {
-			digest.reset();
-		}
-		return result;
+		return resetAndRepeatDigest(result);
 	}
 
 	/**
@@ -233,29 +288,25 @@ public class Digester extends BouncyCastleSupport {
 	 * @param data {@link InputStream} 数据流
 	 * @param bufferLength 缓存长度，不足1使用 {@link IoUtil#DEFAULT_BUFFER_SIZE} 做为默认值
 	 * @return 摘要bytes
-	 * @throws CryptoException IO异常
+	 * @throws IORuntimeException IO异常
 	 */
-	public byte[] digest(InputStream data, int bufferLength) throws CryptoException {
+	public byte[] digest(InputStream data, int bufferLength) throws IORuntimeException {
 		if (bufferLength < 1) {
 			bufferLength = IoUtil.DEFAULT_BUFFER_SIZE;
 		}
-		byte[] buffer = new byte[bufferLength];
-
-		byte[] result = null;
+		
+		byte[] result;
 		try {
-			int read = data.read(buffer, 0, bufferLength);
-
-			while (read > -1) {
-				digest.update(buffer, 0, read);
-				read = data.read(buffer, 0, bufferLength);
+			if (ArrayUtil.isEmpty(this.salt)) {
+				result = digestWithoutSalt(data, bufferLength);
+			} else {
+				result = digestWithSalt(data, bufferLength);
 			}
-			result = digest.digest();
 		} catch (IOException e) {
-			throw new CryptoException(e);
-		} finally {
-			digest.reset();
+			throw new IORuntimeException(e);
 		}
-		return result;
+		
+		return resetAndRepeatDigest(result);
 	}
 
 	/**
@@ -288,4 +339,95 @@ public class Digester extends BouncyCastleSupport {
 	public int getDigestLength() {
 		return this.digest.getDigestLength();
 	}
+
+	// -------------------------------------------------------------------------------- Private method start
+	/**
+	 * 生成摘要
+	 * 
+	 * @param data {@link InputStream} 数据流
+	 * @param bufferLength 缓存长度，不足1使用 {@link IoUtil#DEFAULT_BUFFER_SIZE} 做为默认值
+	 * @return 摘要bytes
+	 * @throws IOException 从流中读取数据引发的IO异常
+	 */
+	private byte[] digestWithoutSalt(InputStream data, int bufferLength) throws IOException {
+		final byte[] buffer = new byte[bufferLength];
+		int read;
+		while ((read = data.read(buffer, 0, bufferLength)) > -1) {
+			this.digest.update(buffer, 0, read);
+		}
+		return this.digest.digest();
+	}
+
+	/**
+	 * 生成摘要
+	 * 
+	 * @param data {@link InputStream} 数据流
+	 * @param bufferLength 缓存长度，不足1使用 {@link IoUtil#DEFAULT_BUFFER_SIZE} 做为默认值
+	 * @return 摘要bytes
+	 * @throws IOException 从流中读取数据引发的IO异常
+	 */
+	private byte[] digestWithSalt(InputStream data, int bufferLength) throws IOException {
+		if (this.saltPosition <= 0) {
+			// 加盐在开头
+			this.digest.update(this.salt);
+		}
+
+		final byte[] buffer = new byte[bufferLength];
+		int total = 0;
+		int read;
+		while ((read = data.read(buffer, 0, bufferLength)) > -1) {
+			total += read;
+			if (this.saltPosition > 0 && total >= this.saltPosition) {
+				if (total != this.saltPosition) {
+					digest.update(buffer, 0, total - this.saltPosition);
+				}
+				// 加盐在中间
+				this.digest.update(this.salt);
+				this.digest.update(buffer, total - this.saltPosition, read);
+			} else {
+				this.digest.update(buffer, 0, read);
+			}
+		}
+
+		if (total < this.saltPosition) {
+			// 加盐在末尾
+			this.digest.update(this.salt);
+		}
+
+		return this.digest.digest();
+	}
+
+	/**
+	 * 生成摘要
+	 * 
+	 * @param datas 数据bytes
+	 * @return 摘要bytes
+	 * @since 4.4.3
+	 */
+	private byte[] doDigest(byte[]... datas) {
+		for (byte[] data : datas) {
+			if (null != data) {
+				this.digest.update(data);
+			}
+		}
+		return this.digest.digest();
+	}
+
+	/**
+	 * 重复计算摘要，取决于{@link #digestCount} 值<br>
+	 * 每次计算摘要前都会重置{@link #digest}
+	 * 
+	 * @param digestData 第一次摘要过的数据
+	 * @return 摘要
+	 */
+	private byte[] resetAndRepeatDigest(byte[] digestData) {
+		final int digestCount = Math.max(1, this.digestCount);
+		reset();
+		for (int i = 0; i < digestCount - 1; i++) {
+			digestData = doDigest(digestData);
+			reset();
+		}
+		return digestData;
+	}
+	// -------------------------------------------------------------------------------- Private method end
 }