+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * 【socket】     socket模块加入到all中
 * 【core】        增加Jdk8DateConverter用于支持jdk8中的时间（issue#IS32N@Gitee）
 * 【core】        StrUtil.subPreGbk优化代码规范（pull#277@Github）
+* 【crypto】      MD5支持16位值生成
+* 【crypto】      Digester支持自定义盐所在位置
 
 ### Bug修复